+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 # v2.25.3
 
-*
+* CacheDict will now accept time-like strings